+++ core/src/oned/ODDXFilmEdgeReader.cpp
@@ -175,7 +175,7 @@ Result DXFilmEdgeReader::decodePattern(int rowNumber, PatternView& next, std::un
 	BitArray dataBits;
 	while (next.isValid(1) && dataBits.size() < clock->dataLength()) {
 
-		int modules = next[0] / clock->moduleSize() + 0.5;
+		int modules = int(next[0] / clock->moduleSize() + 0.5);
 		// even index means we are at a bar, otherwise at a space
 		dataBits.appendBits(next.index() % 2 == 0 ? 0xFFFFFFFF : 0x0, modules);
